+++ app/controllers/profiles_controller.rb
@@ -1,5 +1,5 @@
 class ProfilesController < ApplicationController
   def show
-    # @user = current_user
+    @user = current_user
   end
 end
