+++ app/javascript/packs/application.js
@@ -15,3 +15,40 @@ require("channels")
 //
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
+
+import $ from 'jquery'
+import axios from 'modules/axios'
+import {
+  listenInactiveHeartEvent,
+  listenActiveHeartEvent
+} from 'modules/handle_heart'
+
+const handleHeartDisplay = (hasLiked, albumId) => {
+  if (hasLiked) {
+    $(`#active-heart-icon-${albumId}`).removeClass('hidden')
+  } else {
+    $(`#inactive-heart-icon-${albumId}`).removeClass('hidden')
+  }
+}
+
+document.addEventListener('turbolinks:load', () => {
+  // すべてのアルバムカード要素を取得
+  const albums = document.querySelectorAll('.card')
+
+  albums.forEach((album) => {
+    // 各アルバムのデータセットを取得
+    const dataset = album.dataset
+    const albumId = dataset.albumId
+    if (albumId) {
+      // APIリクエストを送信してlike状態を取得
+      axios.get(`/api/albums/${albumId}/like`)
+        .then((response) => {
+          const hasLiked = response.data.hasLiked
+          handleHeartDisplay(hasLiked, albumId)
+        })
+
+      listenInactiveHeartEvent(albumId)
+      listenActiveHeartEvent(albumId)
+      }
+  })
+})
