+++ app/views/albums/index.html.haml
@@ -37,6 +37,3 @@
           .card_detail_title_author
             %p= album.user.account
           %p= album.title
-
--# = javascript_pack_tag 'album'
-= javascript_pack_tag 'album_index'
